License check

See merge request !17

+++ SmartFare/inc/SmartFareData.h
@@ -1,8 +1,9 @@
 /*
  * SmartFareData.h
  *
- *  Created on: 26 de set de 2016
- *      Author: luisfg30
+ *  Created on: 2016
+ *      Author: Luis Fernando Guerreiro
+ *      		Diego Gabriel Lee
  */
 
 #ifndef INC_SMARTFAREDATA_H_
@@ -19,9 +20,9 @@
 typedef struct {
 	unsigned int userId; // the PICC used stores a 4 byte user ID
 	unsigned int vehicleId;
-	uint8_t fare;
-	uint8_t balance;  // the user balance in cents, may be negative
-	uint8_t distance; // last travel fare calculated
+	unsigned int fare;
+	int balance;  // the user balance in cents, may be negative
+	unsigned int distance; // last travel fare calculated
 	unsigned int
 		inOdometerMeasure; // vehicle odometer read when user get on board
 	RTC_TIME_T inTimestamp;
